deal with some cast warnings, looking into class loading issues with modes

+++ app/src/processing/app/windows/Registry.java
@@ -93,7 +93,7 @@ public class Registry {
    * @return int
    */
   private static int convertBufferToInt(byte[] buf) {
-   return(((int)(buf[0] & 0xff)) + (((int)(buf[1] & 0xff)) << 8) + (((int)(buf[2] & 0xff)) << 16) + (((int)(buf[3] & 0xff)) << 24));
+    return((buf[0] & 0xff) + ((buf[1] & 0xff) << 8) + ((buf[2] & 0xff) << 16) + ((buf[3] & 0xff) << 24));
   }
 
   /**
